add user domain logics, tests
add custom jpql query to prevent N+1
add interest initializer to set informations
add test fixtures to enhance test codes/environments
fix referencing error between N:N relationship(user-userInterest-interest)
add settings/logics for jwt/oauth/cookie based authentications
fix jwtAuthenticationFilter, securityConfig multiple times to find appropriate strategy and enhance both performance and UX

+++ backend/src/main/java/com/ssup/backend/domain/post/PostService.java
@@ -5,19 +5,24 @@ import com.ssup.backend.domain.post.dto.*;
 import com.ssup.backend.domain.post.sort.PostSliceFetcher;
 import com.ssup.backend.domain.post.sort.PostSortType;
 import com.ssup.backend.domain.user.User;
-import com.ssup.backend.domain.user.UserService;
+import com.ssup.backend.domain.user.UserRepository;
+import com.ssup.backend.domain.user.UserStatus;
+import com.ssup.backend.global.exception.ErrorCode;
 import com.ssup.backend.global.exception.SsupException;
+import com.ssup.backend.infra.aop.CheckUserStatus;
 import com.ssup.backend.infra.s3.ImageStorage;
 import com.ssup.backend.infra.s3.ImageType;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.transaction.support.TransactionSynchronizationManager;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
 
+import static com.ssup.backend.domain.post.PostValidator.*;
 import static com.ssup.backend.global.exception.ErrorCode.*;
 
 @Service
@@ -29,26 +34,30 @@ public class PostService {
     private final PostRepository postRepository;
     private final PostSliceFetcher postSliceFetcher;
     private final PostHeartRepository postHeartRepository;
-    private final UserService userService;
+    private final UserRepository userRepository;
     private final ImageStorage imageStorage;
 
-    //todo: userId -> appUser
+    @CheckUserStatus(UserStatus.ACTIVE)
     public PostCreateResponse create(Long userId, List<MultipartFile> images, PostCreateRequest request) {
         Post post = request.toEntity();
         List<String> imageUrls = imageStorage.uploadMultiple(ImageType.POST, images);
         post.addImages(imageUrls);
 
-        User user = userService.findUserById(userId);
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new SsupException(USER_NOT_FOUND));
+
         post.setAuthor(user);
 
         Post savedPost = postRepository.save(post);
         return PostCreateResponse.of(user, savedPost);
     }
 
+    @CheckUserStatus(UserStatus.ACTIVE)
     public PostUpdateResponse update(Long userId, Long postId, List<MultipartFile> addedImages, PostUpdateRequest request) {
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new SsupException(POST_NOT_FOUND));
 
+        validatePostOwner(userId, post.getAuthor().getId());
         updateImages(addedImages, request.getKeepImageUrls(), post);
         post.update(request.getTitle(), request.getContent(), request.getUsingLanguage(), request.getLearningLanguage());
 
@@ -65,7 +74,7 @@ public class PostService {
         return postSliceFetcher.fetch(userId, sort, cursorKey, cursorId, size);
     }
 
-    public PostResponse find(Long id, Long userId) {
+    public PostResponse find(Long userId, Long id) {
         Post post = postRepository.findById(id)
                 .orElseThrow(() -> new SsupException(POST_NOT_FOUND));
 
@@ -81,16 +90,19 @@ public class PostService {
         return PostResponse.of(author, post, heartedByMe);
     }
 
-    public void delete(Long id) {
-        postRepository.deleteById(id);
+    @CheckUserStatus(UserStatus.ACTIVE)
+    public void delete(Long userId, Long id) {
+        Post post = postRepository.findById(id)
+                .orElseThrow(() -> new SsupException(POST_NOT_FOUND));
 
-        //todo: 연관 comment, heart, images 모두 delete
+        validatePostOwner(userId, post.getAuthor().getId());
+        postRepository.deleteById(id);
     }
 
     //=====
 
     private void updateImages(List<MultipartFile> addedImages, List<String> keepImageUrls, Post post) {
-        //삭제 대상 이미지 계산
+        //삭제 대상 이미지 추출
         List<String> removedImages = post.getImageUrls().stream()
                 .filter(url -> !keepImageUrls.contains(url))
                 .toList();
